requirements.txt
Levenshtein==0.26.1
pillow==11.1.0
PyMuPDF==1.25.3
RapidFuzz==3.12.1
